pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ecofuture-preproc"
version = "0.1.0"
description = ""
authors = ["Damien Mannion <damien.mannion@unimelb.edu.au>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9,<3.12"
python-dateutil = "^2.8.2"
geojson = "^3.0.1"
rasterio = "^1.3.6"
rioxarray = "^0.14.1"
pyproj = "^3.5.0"
shapely = "^2.0.1"
requests = "^2.31.0"
boto3 = "^1.26.161"
netcdf4 = "^1.6.4"
tqdm = "^4.65.0"
pyyaml = "^6.0"
dask = "^2023.8.0"
odc-geo = "^0.4.1"
fiona = "^1.9.4.post1"
siphon = "^0.9"
botocore = "^1.31.35"
numpy = "^1.25.2"
xarray = "^2023.8.0"
polars = "^0.18.15"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "23.7.0"
mypy = "^1.3.0"
ruff = "^0.0.285"
types-tqdm = "^4.65.0.1"
boto3-stubs = {extras = ["s3"], version = "^1.26.161"}
types-requests = "^2.31.0.1"
types-pyyaml = "^6.0.12.10"

[tool.poetry.group.interactive.dependencies]
ipython = "^8.14.0"
jupyter = "^1.0.0"
pyqt5 = "^5.15.9"
matplotlib = "^3.7.2"

[tool.poetry.group.dev.dependencies]
fawltydeps = "^0.12.1"

[tool.poetry.scripts]
ecofuture_preproc = "ecofuture_preproc.cli:main"

[tool.mypy]
plugins = "numpy.typing.mypy_plugin"
# corresponding to --strict
warn_unused_configs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_reexport = true
strict_equality = true
extra_checks = true
# except for this
allow_redefinition = true
follow_imports = "skip"

[[tool.mypy.overrides]]
module = [
        "geojson",
        "shapely",
        "rasterio",
	"siphon",
	"siphon.catalog",
	"affine",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "rioxarray"
implicit_reexport = true

[tool.black]
line-length = 88
preview = false
skip-source-first-line = false
skip-string-normalization = false
skip-magic-trailing-comma = false

[tool.ruff]
select = [
        "E",  # error
        "F",  # pyflakes
        "UP",  # pyupgrade
        "N",  # pep8-naming
        "ANN",  # flake8-annotations
        "A",  # flake8-builtins
        "PTH",  # flake8-usepathlib
        "B",  # flake8-bugbear
        "Q",  # flake8-quotes
        "ARG",  # flake8-unused-arguments
        "NPY",  # numpy
        "PERF",  # perflint
        "RUF",  # ruff
]
target-version = "py39"
ignore = [
        "ANN101",  # typing self
]

[tool.ruff.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

[tool.fawltydeps]
ignore_unused = ["black", "fawltydeps", "mypy", "ruff", "ipython"]
output_format = "human_detailed"
